requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
drf-yasg>=1.21,<2.0
django-filter>=23.0,<24.0
mysqlclient>=2.2,<3.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
